% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Activity_DailyData}
\alias{Activity_DailyData}
\title{Activity daily data}
\format{
a dataframe of lots of variables
}
\usage{
Activity_DailyData
}
\description{
Activity daily data
}
\keyword{datasets}
