% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{halo_body_df}
\alias{halo_body_df}
\title{Halo Body Composition}
\usage{
halo_body_df(pathname = getwd())
}
\arguments{
\item{pathname}{path to the Amazon Halo toplevel directory}
}
\description{
Return fat percentage
}
