% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{halo_daily_df}
\alias{halo_daily_df}
\title{Halo daily data dataframe}
\usage{
halo_daily_df(pathname = getwd())
}
\arguments{
\item{pathname}{path to the Amazon Halo toplevel directory}
}
\description{
Return dataframe for all daily data
}
