% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{halo_heartrate_df}
\alias{halo_heartrate_df}
\title{Halo raw heart rate dataframe}
\usage{
halo_heartrate_df(pathname = getwd(), sample = NULL)
}
\arguments{
\item{pathname}{path to the Amazon Halo toplevel directory}

\item{sample}{frequency of rows to include (e.g. 0.2)}
}
\description{
Return dataframe for raw heart rate data
}
