% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{halo_raw_activity_df}
\alias{halo_raw_activity_df}
\title{Halo raw data dataframe}
\usage{
halo_raw_activity_df(pathname = getwd())
}
\arguments{
\item{pathname}{path to the Amazon Halo toplevel directory}
}
\description{
Return dataframe for raw activity data
}
