% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_halo.R
\name{halo_tone_utterances_df}
\alias{halo_tone_utterances_df}
\title{Halo raw tone utterances dataframe}
\usage{
halo_tone_utterances_df(pathname = getwd(), sample = NULL)
}
\arguments{
\item{pathname}{path to the Amazon Halo toplevel directory}
}
\description{
Return dataframe for tone utterances
}
