% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Sleep_Sessions}
\alias{Sleep_Sessions}
\title{Sleep data}
\format{
a dataframe of 5 variables
\describe{
  \item{BedTime}{price, in US dollars}
  \item{`Sleep Start Time`}{weight of the diamond, in carats}
  ...
}
}
\usage{
Sleep_Sessions
}
\description{
Sleep data
}
\keyword{datasets}
